Added social metadata on brew profiles
